+++ src/core/models/fluorescence/fl_1_1_1_2_Model.cpp
@@ -21,6 +21,7 @@
 #include "src/core/equil.h"
 #include "src/core/libmath.h"
 #include "src/core/models.h"
+#include "src/core/thermo.h"
 #include "src/core/toolset.h"
 
 #include <QDebug>
@@ -177,10 +178,10 @@ QSharedPointer<AbstractModel> fl_ItoI_ItoII_Model::Clone()
 
 QString fl_ItoI_ItoII_Model::ParameterComment(int parameter) const
 {
-    if (parameter == 1)
-        return QString("Reaction: AB + A &#8652; A<sub>2</sub>B");
+    if (parameter == 0)
+        return QString("Reaction: A + A &#8652; AB");
     else
-        return QString("Reaction: A + B &#8652; AB");
+        return QString("Reaction: AB + B &#8652; AB<sub>2</sub>");
 }
 
 QString fl_ItoI_ItoII_Model::ModelInfo() const
@@ -192,46 +193,13 @@ QString fl_ItoI_ItoII_Model::ModelInfo() const
 
 QString fl_ItoI_ItoII_Model::AnalyseMonteCarlo(const QJsonObject& object, bool forceAll) const
 {
-
-    QString result = AbstractTitrationModel::AnalyseMonteCarlo(object);
+    QString result = AbstractTitrationModel::AnalyseMonteCarlo(object, forceAll);
 
     if (!forceAll)
         return result;
 
-    QStringList models = object["controller"].toObject()["raw"].toObject().keys();
-
-    QList<qreal> s, s_sf;
-
-    for (int i = 0; i < models.size(); ++i) {
-        QJsonObject model = object["controller"].toObject()["raw"].toObject()[models[i]].toObject();
-        qreal logK11 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[0];
-        qreal logK12 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[1];
-
-        s << BC50::ItoI_ItoII_BC50(logK11, logK12) * 1e6;
-        s_sf << BC50::ItoI_ItoII_BC50_SF(logK11, logK12) * 1e6;
-    }
-    std::sort(s.begin(), s.end());
-    std::sort(s_sf.begin(), s_sf.end());
-
-    SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, 95);
-    SupraFit::ConfidenceBar conf_sf = ToolSet::Confidence(s_sf, 95);
-
-    qreal BC50 = BC50::ItoI_ItoII_BC50(GlobalParameter(0), GlobalParameter(1)) * 1e6;
-    qreal BC50_sf = BC50::ItoI_ItoII_BC50_SF(GlobalParameter(0), GlobalParameter(1)) * 1e6;
-
-    qreal conf_dSl = conf.upper - BC50;
-    qreal conf_dSu = BC50 - conf.lower;
-
-    qreal conf_dSl_sf = conf_sf.upper - BC50_sf;
-    qreal conf_dSu_sf = BC50_sf - conf_sf.lower;
-
-    result += tr("<p>BC50 %1 [+%2,-%3] %4M ... ").arg(BC50).arg(conf_dSu).arg(conf_dSl).arg(QChar(956));
-    result += tr("[%1 - %2] %3M</p>").arg(conf.lower).arg(conf.upper).arg(QChar(956));
-
-    result += tr("<p>BC50 (SF) %1 [+%2,-%3] %4M ... ").arg(BC50_sf).arg(conf_dSu_sf).arg(conf_dSl_sf).arg(QChar(956));
-    result += tr("[%1 - %2] %3M</p>").arg(conf_sf.lower).arg(conf_sf.upper).arg(QChar(956));
-
-    return result;
+    QString bc = Thermo::Statistic2BC50_1_2(GlobalParameter(0), GlobalParameter(1), object);
+    return bc + result;
 }
 
 #include "fl_1_1_1_2_Model.moc"