+++ src/core/models/fluorescence/fl_1_1_Model.h
@@ -63,6 +63,8 @@ public:
 
     virtual QString ModelInfo() const override;
 
+    QString AnalyseMonteCarlo(const QJsonObject& object, bool forceAll = false) const;
+
 protected:
     virtual void CalculateVariables() override;
 };