+++ src/core/models/itc/itc_1_1_Model.cpp
@@ -187,4 +187,16 @@ QString itc_ItoI_Model::ParameterComment(int parameter) const
     return QString("Reaction: A + B &#8652; AB");
 }
 
+QString itc_ItoI_Model::AnalyseMonteCarlo(const QJsonObject& object, bool forceAll) const
+{
+
+    QString result = AbstractItcModel::AnalyseMonteCarlo(object, forceAll);
+
+    if (!forceAll)
+        return result;
+
+    QString bc = Thermo::Statistic2BC50_1(GlobalParameter(0), object);
+    return bc + result;
+}
+
 #include "itc_1_1_Model.moc"