+++ src/core/models/itc/itc_1_1_Model.h
@@ -93,6 +93,8 @@ public:
 
     QString ModelInfo() const override;
 
+    QString AnalyseMonteCarlo(const QJsonObject& object, bool forceAll = false) const;
+
 protected:
     virtual void CalculateVariables() override;
 };