+++ src/core/models/titrations/1_1_Model.h
@@ -73,6 +73,8 @@ public:
 
     virtual QString ModelInfo() const override;
 
+    QString AnalyseMonteCarlo(const QJsonObject& object, bool forceAll = false) const;
+
 protected:
     virtual void CalculateVariables() override;
 };