+++ src/core/models/titrations/2_1_1_1_1_2_Model.cpp
@@ -355,40 +355,9 @@ QString IItoI_ItoI_ItoII_Model::AnalyseMonteCarlo(const QJsonObject& object, boo
     if (!forceAll)
         return result;
 
-    QStringList models = object["controller"].toObject()["raw"].toObject().keys();
-    QList<qreal> s, s_sf;
+    QString bc = Thermo::Statistic2BC50_2_2(GlobalParameter(0), GlobalParameter(1), GlobalParameter(2), object);
 
-    for (int i = 0; i < models.size(); ++i) {
-        QJsonObject model = object["controller"].toObject()["raw"].toObject()[models[i]].toObject();
-        qreal logK21 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[0];
-        qreal logK11 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[1];
-        qreal logK12 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[2];
-
-        s << BC50::IItoI_ItoI_ItoII_BC50(logK21, logK11, logK12) * 1e6;
-        s_sf << BC50::IItoI_ItoI_ItoII_BC50_SF(logK21, logK11, logK12) * 1e6;
-    }
-
-    std::sort(s.begin(), s.end());
-    std::sort(s_sf.begin(), s_sf.end());
-
-    SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, 95);
-    SupraFit::ConfidenceBar conf_sf = ToolSet::Confidence(s_sf, 95);
-
-    qreal BC50 = BC50::IItoI_ItoI_ItoII_BC50(GlobalParameter(0), GlobalParameter(1), GlobalParameter(2)) * 1e6;
-    qreal BC50_sf = BC50::IItoI_ItoI_ItoII_BC50_SF(GlobalParameter(0), GlobalParameter(1), GlobalParameter(2)) * 1e6;
-
-    qreal conf_dSl = conf.upper - BC50;
-    qreal conf_dSu = BC50 - conf.lower;
-
-    qreal conf_dSl_sf = conf_sf.upper - BC50_sf;
-    qreal conf_dSu_sf = BC50_sf - conf_sf.lower;
-
-    result += tr("<p>BC50 %1 [+%2,-%3] %4M ... ").arg(BC50).arg(conf_dSu).arg(conf_dSl).arg(QChar(956));
-    result += tr("[%1 - %2] %3M</p>").arg(conf.lower).arg(conf.upper).arg(QChar(956));
-
-    result += tr("<p>BC50 (SF) %1 [+%2,-%3] %4M ... ").arg(BC50_sf).arg(conf_dSu_sf).arg(conf_dSl_sf).arg(QChar(956));
-    result += tr("[%1 - %2] %3M</p>").arg(conf_sf.lower).arg(conf_sf.upper).arg(QChar(956));
-    return result;
+    return bc + result;
 }
 
 #include "2_1_1_1_1_2_Model.moc"