+++ src/core/thermo.cpp
@@ -34,7 +34,7 @@ QString Statistic2Thermo(qreal K, qreal H, qreal T, const QJsonObject& object)
 
     int index_global = 0;
     int index_local = 0;
-
+    qreal error;
     if (!object.isEmpty()) {
         const QStringList keys = object.keys();
         for (int i = 0; i < keys.size(); ++i) {
@@ -43,14 +43,13 @@ QString Statistic2Thermo(qreal K, qreal H, qreal T, const QJsonObject& object)
                 K11u = confidence["upper"].toDouble();
                 K11l = confidence["lower"].toDouble();
                 index_global = object[keys[i]].toObject()["index"].toString().toInt();
-                //   qDebug() << index_global << object[keys[i]].toObject()["index"].toString();
+                error = 100 - confidence["error"].toDouble();
             }
 
             if (qFuzzyCompare(object[keys[i]].toObject()["value"].toDouble(), H)) {
                 dH11u = confidence["upper"].toDouble();
                 dH11l = confidence["lower"].toDouble();
                 index_local = object[keys[i]].toObject()["index"].toString().split("|")[0].toInt();
-                //     qDebug() << index_local << object[keys[i]].toObject()["index"].toString();
             }
         }
     }
@@ -61,16 +60,9 @@ QString Statistic2Thermo(qreal K, qreal H, qreal T, const QJsonObject& object)
     qreal dGl = ToolSet::K2G(K11u, T);
     qreal dGu = ToolSet::K2G(K11l, T);
 
-    qreal conf_dGu = dG - dGu;
+    qreal conf_dGu = dGu - dG;
     qreal conf_dGl = dG - dGl;
 
-    /*
-        qreal dSll = ToolSet::GHE(dGl, dH11l, T);
-        qreal dSul = ToolSet::GHE(dGu, dH11l, T);
-        qreal dSuu = ToolSet::GHE(dGu, dH11u, T);
-        qreal dSlu = ToolSet::GHE(dGl, dH11u, T);
-        */
-    // qDebug() << dSll << dSuu << dSul << dSlu << conf.lower << conf.upper << dS;
 
     result += "<table>";
     result += "<tr><td><b>Complexation Constant K </b></td><td>" + Print::printDouble(qPow(10, K)) + "</td>";
@@ -79,10 +71,15 @@ QString Statistic2Thermo(qreal K, qreal H, qreal T, const QJsonObject& object)
         result += "<td> (+" + Print::printDouble(qPow(10, K11u) - qPow(10, K)) + "/-" + Print::printDouble(qPow(10, K) - qPow(10, K11l)) + ")</td>";
     result += "<td> M</td></tr>";
 
+    if (!object.isEmpty())
+        result += QString("<tr><td><b></b></td><td>[%1 - %2]  </td><td></td><td>M</td></tr>").arg(qPow(10, K11l)).arg(qPow(10, K11u));
+
     result += "<tr><td><b>Free Enthalpy of Complexation &Delta;G </b></td><td>" + Print::printDouble(dG / 1000.0, 3) + "</td>";
     if (!object.isEmpty())
         result += "<td> (+" + Print::printDouble(conf_dGu / 1000, 3) + "/-" + Print::printDouble(conf_dGl / 1000, 3) + ")</td>";
     result += "<td>kJ/mol</td></tr>";
+    if (!object.isEmpty())
+        result += QString("<tr><td><b></b></td><td>[%1 - %2]  </td><td></td><td>kJ/mol</td></tr>").arg(dGl / 1000.0).arg(dGu / 1000.0);
 
     if (!qFuzzyCompare(H, 0)) {
 
@@ -94,32 +91,177 @@ QString Statistic2Thermo(qreal K, qreal H, qreal T, const QJsonObject& object)
             qreal K = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[index_global];
             QVector<qreal> local = ToolSet::String2DoubleVec(model["localParameter"].toObject()["data"].toObject()["0"].toString());
             qreal H = local[index_local];
-            // qDebug() << K << H;
             s << ToolSet::GHE(ToolSet::K2G(K, T), H, T);
         }
 
         std::sort(s.begin(), s.end());
 
-        SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, 95);
+        SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, error);
 
-        qreal conf_dH11u = H - dH11u;
+        qreal conf_dH11u = dH11u - H;
         qreal conf_dH11l = H - dH11l;
 
-        qreal conf_dSl = conf.upper - dS;
-        qreal conf_dSu = dS - conf.lower;
+        qreal conf_dSu = conf.upper - dS;
+        qreal conf_dSl = dS - conf.lower;
 
         result += "<tr><td><b>Enthalpy of Complexation &Delta;H</b></td><td>" + Print::printDouble(H / 1000.0) + "</td>";
+
         if (!object.isEmpty())
-            result += "<td>(" + Print::printDouble(conf_dH11u / 1000, 3) + "/-" + Print::printDouble(conf_dH11l / 1000, 3) + ")</td>";
+            result += "<td>(" + Print::printDouble(conf_dH11u / 1000.0, 3) + "/-" + Print::printDouble(conf_dH11l / 1000.0, 3) + ")</td>";
         result += "<td>kJ/mol</td></tr>";
+
+        if (!object.isEmpty())
+            result += QString("<tr><td><b></b></td><td>[%1 - %2]  </td><td></td><td>J/(molK)</td></tr>").arg(dH11l / 1000.0).arg(dH11u / 1000.0);
 
         result += "<tr><td><b>Entropy of Complexation &Delta;S</b></td><td>" + Print::printDouble(dS) + "</td>";
         if (!object.isEmpty())
             result += "<td>(+" + Print::printDouble(conf_dSu, 3) + "/-" + Print::printDouble(conf_dSl, 3) + ")</td>";
         result += "<td>J/(molK)</td></tr>";
+        if (!object.isEmpty())
+            result += QString("<tr><td><b></b></td><td>[%1 - %2]  </td><td></td><td>J/(molK)</td></tr>").arg(conf.lower).arg(conf.upper);
     }
     result += "</table>";
 
+    return result;
+}
+
+QString Statistic2BC50_1(const qreal logK11, const QJsonObject& object)
+{
+    QStringList models = object["controller"].toObject()["raw"].toObject().keys();
+    qreal error = 100 - object["0"].toObject()["confidence"].toObject()["error"].toDouble();
+
+    QList<qreal> s;
+
+    for (int i = 0; i < models.size(); ++i) {
+        QJsonObject model = object["controller"].toObject()["raw"].toObject()[models[i]].toObject();
+        qreal logK11 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[0];
+        s << BC50::ItoI_BC50(logK11) * 1e6;
+    }
+    std::sort(s.begin(), s.end());
+
+    SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, error);
+    qreal BC50 = BC50::ItoI_BC50(logK11) * 1e6;
+
+    qreal conf_dSl = conf.upper - BC50;
+    qreal conf_dSu = BC50 - conf.lower;
+
+    QString result;
+
+    result += QString("<p>BC50 %1 [+%2,-%3] %4M ... ").arg(BC50).arg(conf_dSl).arg(conf_dSu).arg(QChar(956));
+    result += QString("[%1 - %2] %3M</p>").arg(conf.lower).arg(conf.upper).arg(QChar(956));
+
+    return result;
+}
+
+QString Statistic2BC50_1_2(const qreal logK11, const qreal logK12, const QJsonObject& object)
+{
+    QStringList models = object["controller"].toObject()["raw"].toObject().keys();
+    QList<qreal> s, s_sf;
+    qreal error = 100 - object["0"].toObject()["confidence"].toObject()["error"].toDouble();
+
+    for (int i = 0; i < models.size(); ++i) {
+        QJsonObject model = object["controller"].toObject()["raw"].toObject()[models[i]].toObject();
+        qreal logK11 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[0];
+        qreal logK12 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[1];
+
+        s << BC50::ItoI_ItoII_BC50(logK11, logK12) * 1e6;
+        s_sf << BC50::ItoI_ItoII_BC50_SF(logK11, logK12) * 1e6;
+    }
+
+    std::sort(s.begin(), s.end());
+    std::sort(s_sf.begin(), s_sf.end());
+
+    SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, error);
+    SupraFit::ConfidenceBar conf_sf = ToolSet::Confidence(s_sf, error);
+
+    qreal BC50 = BC50::ItoI_ItoII_BC50(logK11, logK12) * 1e6;
+    qreal BC50_sf = BC50::ItoI_ItoII_BC50_SF(logK11, logK12) * 1e6;
+
+    qreal conf_dSl = conf.upper - BC50;
+    qreal conf_dSu = BC50 - conf.lower;
+
+    qreal conf_dSl_sf = conf_sf.upper - BC50_sf;
+    qreal conf_dSu_sf = BC50_sf - conf_sf.lower;
+
+    QString result;
+
+    result += QString("<p>BC50 %1 [+%2,-%3] %4M ... ").arg(BC50).arg(conf_dSl).arg(conf_dSu).arg(QChar(956));
+    result += QString("[%1 - %2] %3M</p>").arg(conf.lower).arg(conf.upper).arg(QChar(956));
+
+    result += QString("<p>BC50 (SF) %1 [+%2,-%3] %4M ... ").arg(BC50_sf).arg(conf_dSl_sf).arg(conf_dSu_sf).arg(QChar(956));
+    result += QString("[%1 - %2] %3M</p>").arg(conf_sf.lower).arg(conf_sf.upper).arg(QChar(956));
+
+    return result;
+}
+
+QString Statistic2BC50_2_1(const qreal logK21, const qreal logK11, const QJsonObject& object)
+{
+    QString result;
+    QList<qreal> s;
+
+    QStringList models = object["controller"].toObject()["raw"].toObject().keys();
+    qreal error = 100 - object["0"].toObject()["confidence"].toObject()["error"].toDouble();
+
+    for (int i = 0; i < models.size(); ++i) {
+        QJsonObject model = object["controller"].toObject()["raw"].toObject()[models[i]].toObject();
+        qreal logK21 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[0];
+        qreal logK11 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[1];
+
+        s << BC50::IItoI_ItoI_BC50(logK21, logK11) * 1e6;
+    }
+
+    std::sort(s.begin(), s.end());
+
+    SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, error);
+    qreal BC50 = BC50::IItoI_ItoI_BC50(logK21, logK11) * 1e6;
+
+    qreal conf_dSl = conf.upper - BC50;
+    qreal conf_dSu = BC50 - conf.lower;
+
+    result += QString("<p>BC50 %1 [+%2,-%3] %4M ... ").arg(BC50).arg(conf_dSl).arg(conf_dSu).arg(QChar(956));
+    result += QString("[%1 - %2] %3M</p>").arg(conf.lower).arg(conf.upper).arg(QChar(956));
+
+    return result;
+}
+
+QString Statistic2BC50_2_2(const qreal logK21, const qreal logK11, const qreal logK12, const QJsonObject& object)
+{
+    QStringList models = object["controller"].toObject()["raw"].toObject().keys();
+    QList<qreal> s, s_sf;
+    qreal error = 100 - object["0"].toObject()["confidence"].toObject()["error"].toDouble();
+
+    for (int i = 0; i < models.size(); ++i) {
+        QJsonObject model = object["controller"].toObject()["raw"].toObject()[models[i]].toObject();
+        qreal logK21 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[0];
+        qreal logK11 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[1];
+        qreal logK12 = ToolSet::String2DoubleVec(model["globalParameter"].toObject()["data"].toObject()["0"].toString())[2];
+
+        s << BC50::IItoI_ItoI_ItoII_BC50(logK21, logK11, logK12) * 1e6;
+        s_sf << BC50::IItoI_ItoI_ItoII_BC50_SF(logK21, logK11, logK12) * 1e6;
+    }
+
+    std::sort(s.begin(), s.end());
+    std::sort(s_sf.begin(), s_sf.end());
+
+    SupraFit::ConfidenceBar conf = ToolSet::Confidence(s, error);
+    SupraFit::ConfidenceBar conf_sf = ToolSet::Confidence(s_sf, error);
+
+    qreal BC50 = BC50::IItoI_ItoI_ItoII_BC50(logK21, logK11, logK12) * 1e6;
+    qreal BC50_sf = BC50::IItoI_ItoI_ItoII_BC50_SF(logK21, logK11, logK12) * 1e6;
+
+    qreal conf_dSl = conf.upper - BC50;
+    qreal conf_dSu = BC50 - conf.lower;
+
+    qreal conf_dSl_sf = conf_sf.upper - BC50_sf;
+    qreal conf_dSu_sf = BC50_sf - conf_sf.lower;
+
+    QString result;
+
+    result += QString("<p>BC50 %1 [+%2,-%3] %4M ... ").arg(BC50).arg(conf_dSl).arg(conf_dSu).arg(QChar(956));
+    result += QString("[%1 - %2] %3M</p>").arg(conf.lower).arg(conf.upper).arg(QChar(956));
+
+    result += QString("<p>BC50 (SF) %1 [+%2,-%3] %4M ... ").arg(BC50_sf).arg(conf_dSl_sf).arg(conf_dSu_sf).arg(QChar(956));
+    result += QString("[%1 - %2] %3M</p>").arg(conf_sf.lower).arg(conf_sf.upper).arg(QChar(956));
     return result;
 }
 }